# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SHEild AI 2.0 API Contract
  version: 0.1.0-mvp
  description: |
    SHEild AI 2.0 Phase 1 MVP Contract (Women-focused safety MVP).
    
    Architectural & Safety Guarantees:
    - Base URL: /api/v1
    - Auth: Bearer Supabase JWT (ES256/RS256 verified against JWKS)
    - Incident status and Notification/Alert status are strictly separated.
    - FCM push accepted != delivered != acknowledged. Provider acceptance only proves provider queued the notification.
    - Designed to be demographic-neutral so children, elderly, and other groups can be supported without breaking changes.
    - College minor project: zero paid service dependencies assumed.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development backend
paths:
...
  /guardians:
    get:
      summary: List user's accepted and pending guardians
      description: Returns the user's active trusted guardians along with priority escalation sequence.
      operationId: listGuardians
      tags:
        - Guardians
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of linked guardians.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianRelationshipList' }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /guardians/invites:
    get:
      summary: List the owner's invitation history
      description: Invitation hashes are never returned.
      operationId: listGuardianInvites
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Invitations created by this owner. Token hashes are excluded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianInviteList' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      summary: Create a one-time guardian invitation
      description: Authenticated owner only. Returns the random token once; only its SHA-256 hash is persisted. Token expires after seven days.
      operationId: createGuardianInvite
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, phone_number, priority]
              properties:
                display_name: { type: string, minLength: 1, maxLength: 120 }
                phone_number: { type: string, minLength: 3, maxLength: 40 }
                priority: { type: integer, minimum: 1, maximum: 2147483647 }
      responses:
        '201':
          description: Invite created with one-time token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianInviteCreated' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '409': { description: Priority collision. }
  /guardians/accept:
    post:
      summary: Accept a guardian invitation
      description: Atomic, one-use consent operation; authenticated user cannot accept their own invitation.
      operationId: acceptGuardianInvite
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 32 }
      responses:
        '201':
          description: Accepted guardian relationship.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianRelationship' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '409': { description: Invalid, expired, used, duplicate, or priority-conflicting invite. }
  /guardians/relationships:
    get:
      summary: List relationships where the caller is the guardian
      description: RLS exposes only relationships linked to the authenticated guardian account.
      operationId: listMyGuardianships
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Authorized incoming guardian relationships.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianRelationshipList' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /guardians/{link_id}/revoke:
    post:
      summary: Revoke an accepted guardian relationship
      description: Owner only; revocation preserves the relationship row and immediately removes relationship-based RLS visibility.
      operationId: revokeGuardian
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: link_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Revoked relationship.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianRelationship' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /guardians/{link_id}/priority:
    patch:
      summary: Update guardian priority
      description: Owner only; active priority uniqueness is enforced by the database.
      operationId: updateGuardianPriority
      tags: [Guardians]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: link_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority]
              properties: { priority: { type: integer, minimum: 1, maximum: 2147483647 } }
      responses:
        '200':
          description: Updated relationship.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuardianRelationship' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { description: Active guardian priority collision. }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase user access token.
  responses:
...
    UnauthorizedError:
      description: Missing or invalid Supabase Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Token has expired or is invalid."
              request_id: "req_01j8k9y"
    ForbiddenError:
      description: Access denied. User is not the owner or an authorized guardian.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "FORBIDDEN"
              message: "You are not authorized to view or acknowledge this incident."
              request_id: "req_01j8k9z"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: "NOT_FOUND"
              message: "Requested incident or journey does not exist."
              request_id: "req_01j8k90"
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        detail:
          type: string
          description: Compatibility message retained for existing clients.
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "BAD_REQUEST"
            message:
              type: string
              example: "Detailed error description."
            request_id:
              type: string
              nullable: true
              example: "req_01j8k9x"
...
    GuardianRelationship:
      type: object
      required: [id, owner_id, guardian_id, priority, status]
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        guardian_id: { type: string, format: uuid }
        display_name: { type: string, nullable: true }
        phone_number: { type: string, nullable: true, description: Masked for owner guardian-list responses. }
        priority: { type: integer, minimum: 1 }
        status: { type: string, enum: [invited, accepted, revoked] }
        consented_at: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true }
    GuardianRelationshipList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/GuardianRelationship' }
    GuardianInviteCreated:
      type: object
      required: [id, display_name, priority, expires_at, token]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        priority: { type: integer, minimum: 1 }
        expires_at: { type: string, format: date-time }
        token: { type: string, description: Plaintext invitation token returned once to the owner. }
    GuardianInviteItem:
      type: object
      required: [id, display_name, priority, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        phone_number: { type: string }
        priority: { type: integer, minimum: 1 }
        expires_at: { type: string, format: date-time }
        accepted_by: { type: string, format: uuid, nullable: true }
        used_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    GuardianInviteList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/GuardianInviteItem' }
...